[Changes] refactoring code, changing structure
