+++ src/App.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import { Route, Routes, BrowserRouter as Router } from "react-router-dom";
-import { GlobalStyles } from "./components/styles/GlobalStyles";
+import { GlobalStyles } from "./styles/GlobalStyles";
 import MainPage from "./Pages/MainPage";
 import NewsBoard from "./Pages/NewsBlock/NewsBoard/index";
