+++ src/Pages/NewsBlock/NewsPiece/NewsPiece.tsx
@@ -10,12 +10,12 @@ import {
 } from "../NewsPiece/styled";
 import MiddleImage from "../../../assets/NewsBoardImages/Середина.png";
 
-interface NewsPieceBlockProps {
+type NewsPieceBlockProps = {
   image: string;
   title: string;
   description: string;
   date: string;
-}
+};
 
 function NewsPieceBlock({ image, title, description, date }: NewsPieceBlockProps) {
   return (
