+++ src/components/SVG/SvgStyles.tsx
