+++ src/components/SVG/TelegramSVG.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import { SvgWrapper } from "../styles/SvgStyles";
+import { SvgWrapper } from "./SvgStyles";
 function InstagramSVG() {
   return (
     <SvgWrapper>
