+++ src/components/SVG/YoutubeSVG.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import { SvgWrapper } from "../styles/SvgStyles";
+import { SvgWrapper } from "./SvgStyles";
 function YoutubeSVG() {
   return (
     <SvgWrapper>
